refactor: move constants in hook to another directory
